# server/documentation/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cosmian Authentication Verifier
  version: 1.0.0
  description: |
    High-performance authentication and session management server.
    All endpoints are served over **HTTPS only**.

    ## Authentication for admin endpoints
    Admin endpoints require a session cookie obtained by logging into the `_` realm:
    ```
    POST /login?realm=_
    Authorization: Basic <base64(username:password)>
    ```
    The server responds with `Set-Cookie: _ea_=<value>; HttpOnly; Secure; SameSite=Strict`.
    Pass this cookie on every subsequent admin request.
servers:
  - url: https://localhost:8443
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: _ea_
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mtls:
      type: mutualTLS
    appToken:
      type: apiKey
      in: header
      name: X-Vault-Token
      description: |
        AppRole-compatible bearer token issued by this server's AppRole or
        Kubernetes login endpoints.  Required on `/auth/token/*` endpoints.
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Realm 'unknown' not found

    # ── AppRole Auth API ─────────────────────────────────────────────────────────────
...
    OAuthClientRequest:
      type: object
      required: [client_name, redirect_uris]
      properties:
        client_name:
          type: string
          description: Human-readable client name.
          example: My Web App
        redirect_uris:
          type: array
          items:
            type: string
          description: |
            Allowed redirect URIs. Must be `https` (or `http` on localhost).
            Matched exactly at the authorization endpoint.
          example:
            - https://app.example.com/callback
        grant_types:
          type: array
          items:
            type: string
          description: Allowed grant types.
          default: [authorization_code, refresh_token]
          example: [authorization_code, refresh_token]
        response_types:
          type: array
          items:
            type: string
            enum: [code]
          default: [code]
        scopes:
          type: array
          items:
            type: string
          example: [openid, profile, email, offline_access, roles]
        token_endpoint_auth_method:
          type: string
          enum: [client_secret_basic, client_secret_post, none]
          default: client_secret_basic
          description: |
            `none` designates a public (PKCE-only) client with no secret.
    OAuthClientResponse:
      type: object
      required: [client_id, client_name, redirect_uris, grant_types, response_types, scopes, token_endpoint_auth_method, realm, created_at]
      properties:
        client_id:
          type: string
          example: client-Ab3...
        client_secret:
          type: string
          description: |
            The generated client secret. Present **only** in the response to a
            client-creation request (or when a new secret is issued on update);
            never returned on reads.
        client_name:
          type: string
        redirect_uris:
          type: array
          items:
            type: string
        grant_types:
          type: array
          items:
            type: string
        response_types:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        token_endpoint_auth_method:
          type: string
        realm:
          type: string
        created_at:
          type: integer
          format: int64
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: RFC 9068 `at+jwt` access token.
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access-token lifetime in seconds.
        scope:
          type: string
          description: Space-delimited granted scope.
        id_token:
          type: string
          description: OIDC ID token (present when the `openid` scope was granted).
        refresh_token:
          type: string
          description: Opaque refresh token (present when offline access is granted).
    OAuthError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid_grant
        error_description:
          type: string
  responses:
    NoContent:
      description: Success — empty body.
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not authorised.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  parameters:
    realmId:
      name: realm_id
      in: path
      required: true
      schema:
        type: string
      example: my-service
...
    clientId:
      name: client_id
      in: path
      required: true
      schema:
        type: string
      description: OAuth 2.0 / OIDC client identifier.
      example: client-Ab3xY...
paths:
...
  /oidc/token:
    post:
      tags: [OIDC]
      summary: Token endpoint
      operationId: oidcToken
      description: |
        Issues tokens for the `authorization_code` (with mandatory PKCE),
        `refresh_token` (rotating, with reuse detection), and
        `client_credentials` grants. Client authentication via
        `client_secret_basic` or `client_secret_post` (public clients use PKCE).
      security:
        - basicAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token, client_credentials]
                code: {type: string}
                redirect_uri: {type: string}
                code_verifier: {type: string}
                refresh_token: {type: string}
                scope: {type: string}
                client_id: {type: string}
                client_secret: {type: string}
      responses:
        '200':
          description: Token response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: OAuth error (invalid_request, invalid_grant, ...).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          description: Invalid client authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /oidc/userinfo:
    get:
      tags: [OIDC]
      summary: UserInfo endpoint
      operationId: oidcUserinfoGet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scope-gated claims for the token subject.
          content:
            application/json:
              schema: {type: object}
        '401':
          description: Missing or invalid access token.
    post:
      tags: [OIDC]
      summary: UserInfo endpoint (POST)
      operationId: oidcUserinfoPost
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scope-gated claims for the token subject.
          content:
            application/json:
              schema: {type: object}
        '401':
          description: Missing or invalid access token.
  /oidc/introspect:
    post:
      tags: [OIDC]
      summary: Token introspection (RFC 7662)
      operationId: oidcIntrospect
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
                token_type_hint: {type: string}
      responses:
        '200':
          description: Introspection response (`active` plus metadata when active).
          content:
            application/json:
              schema: {type: object}
  /oidc/revoke:
    post:
      tags: [OIDC]
      summary: Token revocation (RFC 7009)
      operationId: oidcRevoke
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
                token_type_hint: {type: string}
      responses:
        '200':
          description: Always 200 OK, regardless of whether the token existed.
  /realms/{realm_id}/clients:
    post:
      tags: [OIDC Clients]
      summary: Register an OAuth client
      operationId: createOAuthClient
      parameters:
        - $ref: '#/components/parameters/realmId'
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthClientRequest'
      responses:
        '201':
          description: Created client (secret returned once for confidential clients).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [OIDC Clients]
      summary: List OAuth clients in a realm
      operationId: listOAuthClients
      parameters:
        - $ref: '#/components/parameters/realmId'
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of clients (secrets omitted).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OAuthClientResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /realms/{realm_id}/clients/{client_id}:
    get:
      tags: [OIDC Clients]
      summary: Get an OAuth client
      operationId: getOAuthClient
      parameters:
        - $ref: '#/components/parameters/realmId'
        - $ref: '#/components/parameters/clientId'
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The client (secret omitted).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClientResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [OIDC Clients]
      summary: Update an OAuth client
      operationId: updateOAuthClient
      parameters:
        - $ref: '#/components/parameters/realmId'
        - $ref: '#/components/parameters/clientId'
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthClientRequest'
      responses:
        '200':
          description: Updated client.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClientResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [OIDC Clients]
      summary: Delete an OAuth client
      operationId: deleteOAuthClient
      parameters:
        - $ref: '#/components/parameters/realmId'
        - $ref: '#/components/parameters/clientId'
      security:
        - cookieAuth: []
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'